refactor(ui): merge duplicated functions
Fixes #3904
